+++ RELEASE_NOTES.md
@@ -39,6 +39,12 @@ Examples
 * FIXED: Typos in the `complements/restrict`, `elephants`, and `permutations`
 examples `NOTES.md` files that caused errors when open and run as notebooks.
 
+Installers and installation scripts
+-----------------------------------
+
+* FIXED: The Handbook and the release build scripts to work with both BSD and
+GNU versions of the `sed` command.
+
 
 3.88.0 - January 29, 2025
 =========================